# perf: optimize MSM routines